requirements.txt
aiofiles
aiohttp
async-lru
async-timeout
attrs
certifi
chardet
dotmap
flake8
h11
httpcore
httpx
idna
mccabe
multidict
pillow
pyaes
pycodestyle
pyflakes
pyrogram
pytgcalls[pyrogram]
pySocks
python-ARQ
python-dotenv
rfc3986
sniffio
tgcrypto
typing-extensions
yarl
youtube-search-python
zipp
psutil
youtube_search
speedtest-cli
git+https://github.com/yt-dlp/yt-dlp
aioredis
